# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "required": [
          "detail"
        ],
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ModelOverrideIn": {
        "additionalProperties": false,
        "description": "One role's model and/or effort, replacing what the tier would have given it.\n\nBoth fields are optional and independent: naming only an effort keeps the tier's model\nand vice versa. The literal sets are the allowlist itself rather than a copy of it, so\na model this engine refuses is a 422 at the edge and never reaches the launcher — and\n`claude-fable-5`, which the CLI silently downgrades to Opus 5, is refused here by not\nbeing in the set (the launcher's error explains why if it arrives another way).\n\nA row may name **either provider's** model regardless of the run's `provider`: mixing\nproviders within one run is the point of the field, not an accident of it.\n\n`effort` is checked here only against the engine's whole vocabulary. Whether a given\nlevel exists on the chosen *model* is the model's own business — every model publishes\nits own ladder in `/api/capabilities`, and a level outside it is clamped, visibly, in\nthe resolved table. **A client offering an effort picker must read the ladder off the\nselected model, not off its provider.**",
        "properties": {
          "effort": {
            "anyOf": [
              {
                "enum": [
                  "low",
                  "medium",
                  "high",
                  "xhigh",
                  "max"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effort"
          },
          "model": {
            "anyOf": [
              {
                "enum": [
                  "claude-opus-5",
                  "fable",
                  "gpt-5.6-luna",
                  "gpt-5.6-sol",
                  "gpt-6-astra"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          }
        },
        "title": "ModelOverrideIn",
        "type": "object"
      },
      "ModelOverrideSettingIn": {
        "additionalProperties": false,
        "description": "One role's model and/or effort on the **settings** body. Deliberately plain strings.\n\nNot `schemas.launch.ModelOverrideIn`, which types both fields as the literal allowlist.\nThat is right for a launch — a run must never reach the launcher naming a model this\nengine refuses — and wrong here, because this endpoint's entire contract is that a\ndefault which could not become a run comes back carrying the *model policy's own*\nsentence. A literal union answers `claude-fable-5` with \"Input should be\n'claude-opus-5', 'fable', 'gpt-5.6-luna' or 'gpt-5.6-sol'\", which is true and useless:\nthe fact worth telling anybody is that the CLI accepts that id and silently runs Opus 5,\nand it lives in `MODEL_TRAPS`, one layer below where Pydantic stops.\n\nSo the values go through as strings and `validate_model`/`validate_effort` refuse them\nwith a reason. `extra=\"forbid\"` stays: an override carries these two fields and nothing\nelse, and an invented field is a malformed request rather than a policy question.",
        "properties": {
          "effort": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effort"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          }
        },
        "title": "ModelOverrideSettingIn",
        "type": "object"
      },
      "ModelSettingsResponse": {
        "additionalProperties": false,
        "description": "The stored default and the table it resolves to.\n\n`source` is `stored` once the editor has saved anything and `built_in` before that, so a\nUI can tell \"this is what the system does\" from \"this is what somebody chose\".\n\n`tiers` is the stored per-tier buckets and `overrides` is a **read-only mirror** of\n`tiers[tier]` — the same object the launcher and the workshop read, served so a client that\nonly cares about the tier in force does not have to index the buckets itself. It is derived\non every response rather than stored, so the mirror cannot fall out of step with what it\nmirrors. `table` is the resolved table for the **selected** tier.",
        "properties": {
          "overrides": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ModelOverrideIn"
            },
            "title": "Overrides",
            "type": "object"
          },
          "provider": {
            "$ref": "#/components/schemas/Provider"
          },
          "source": {
            "enum": [
              "stored",
              "built_in"
            ],
            "title": "Source",
            "type": "string"
          },
          "table": {
            "items": {
              "$ref": "#/components/schemas/ModelSettingsRow"
            },
            "title": "Table",
            "type": "array"
          },
          "tier": {
            "$ref": "#/components/schemas/ModelTier"
          },
          "tiers": {
            "additionalProperties": {
              "additionalProperties": {
                "$ref": "#/components/schemas/ModelOverrideIn"
              },
              "type": "object"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/ModelTier"
            },
            "title": "Tiers",
            "type": "object"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "overrides",
          "provider",
          "source",
          "table",
          "tier",
          "tiers",
          "updated_at"
        ],
        "title": "ModelSettingsResponse",
        "type": "object"
      },
      "ModelSettingsRow": {
        "additionalProperties": false,
        "description": "One resolved step of the default table.",
        "properties": {
          "effort": {
            "enum": [
              "low",
              "medium",
              "high",
              "xhigh",
              "max"
            ],
            "title": "Effort",
            "type": "string"
          },
          "grounded": {
            "title": "Grounded",
            "type": "boolean"
          },
          "model": {
            "enum": [
              "claude-opus-5",
              "fable",
              "gpt-5.6-luna",
              "gpt-5.6-sol",
              "gpt-6-astra"
            ],
            "title": "Model",
            "type": "string"
          },
          "model_class": {
            "enum": [
              "heavy",
              "light"
            ],
            "title": "Model Class",
            "type": "string"
          },
          "note": {
            "title": "Note",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "effort",
          "grounded",
          "model",
          "model_class",
          "note",
          "role"
        ],
        "title": "ModelSettingsRow",
        "type": "object"
      },
      "ModelSettingsUpdate": {
        "additionalProperties": false,
        "description": "`PUT` body. Every field optional; an absent one keeps the built-in default's value.\n\n`overrides` may name either provider's model on any row — mixing providers inside one\ntable is the point of the field. The effort named for a row is checked against the\nengine's vocabulary in the service and against **the chosen model's own ladder** when the\ntable is resolved, which is where a level the model does not have gets clamped, visibly,\ninto the table this endpoint returns.\n\nProvider, tier and the override role keys are strings for the same reason the two fields\nof `ModelOverrideSettingIn` are: every one of them has an engine validator whose refusal\nnames the value *and* the alternatives, and a wire enum in front of that validator\nreplaces the message with a list. Tightening any of these back into an enum silently\nturns this endpoint's 400s into generic 422s again.",
        "properties": {
          "overrides": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ModelOverrideSettingIn"
            },
            "title": "Overrides",
            "type": "object"
          },
          "provider": {
            "default": "anthropic",
            "title": "Provider",
            "type": "string"
          },
          "tier": {
            "default": "max",
            "title": "Tier",
            "type": "string"
          },
          "tiers": {
            "additionalProperties": {
              "additionalProperties": {
                "$ref": "#/components/schemas/ModelOverrideSettingIn"
              },
              "type": "object"
            },
            "title": "Tiers",
            "type": "object"
          }
        },
        "title": "ModelSettingsUpdate",
        "type": "object"
      },
...
      "ModelTier": {
        "description": "A preset that moves thinking effort — and, for `low` alone, the model too.\n\nFor `max`, `high` and `med`, which model a step runs on is fixed by how hard the step is\n— heavy steps take the provider's frontier model, light steps take its workhorse — so\n`med` is a run that thinks less, never a run that thinks with something weaker.\nClustering is pinned to low effort in every tier.\n\n`low` is the speed preset and the one exception: it pins **every** step to `gpt-5.6-luna`\nat high effort whatever the run's provider says, so a Low run drives the Codex CLI and\nneeds it installed. That is published rather than implied — `/api/capabilities` carries\nthe pin and the CLI requirement per tier in `models.tier_catalog` — because a tier\nchanging models is only a problem when nobody can see it.\n\nThe older names `standard`/`maximum` and the pre-floor `balanced`/`quality` are gone\nfrom the wire but are still accepted on the way in and mapped by\n`engine/runners.normalise_tier`, because stored run configs say them.",
        "enum": [
          "max",
          "high",
          "med",
          "low"
        ],
        "title": "ModelTier",
        "type": "string"
      },
...
      "Provider": {
        "description": "Whose CLI runs a step. A quick-set for the whole table, never a lock on one row:\nsteps mix providers freely, and the runner for a call is chosen from that row's model.",
        "enum": [
          "anthropic",
          "openai"
        ],
        "title": "Provider",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "ctx",
          "input",
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local-first API for Oracle run orchestration and artifact discovery.",
    "title": "Oracle GUI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/settings/models": {
      "get": {
        "description": "The default every new run and workshop inherits when it does not state its own.",
        "operationId": "get_model_settings_api_settings_models_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Model Settings",
        "tags": [
          "settings"
        ]
      },
      "put": {
        "description": "Replace the default. Refuses anything that could not become a run.\n\nA whole-value replace rather than a patch: the editor always holds the complete table,\nand a partial update would make \"what is the default\" depend on the order two tabs saved\nin. `tiers` is part of that value — a save is every preset the editor is holding, not the\nselected one alone.\n\nEvery refusal is `400 model_policy` carrying the engine's own sentence — the model trap,\nthe allowlist, the effort rungs, the role names, the providers this build can run. The\n`except` below is the endpoint's contract, not a safety net, which is why the body's\nfields are strings: an enum in front of it would make this branch unreachable and hand\nthe client a 422 that names alternatives without naming the problem.",
        "operationId": "put_model_settings_api_settings_models_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelSettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Model Settings",
        "tags": [
          "settings"
        ]
      }
    },
...
  }
}